Decouple form save code from async task
